+++ kahuna/public/js/upload/recent/recent-uploads.html
@@ -1,0 +1,14 @@
+<p ng:if="! ctrl.myUploads">
+    Loading…
+</p>
+<p ng:if="ctrl.myUploads.data.length == 0">
+    You haven’t uploaded anything yet.
+</p>
+<ul ng:if="ctrl.myUploads.data.length > 0">
+    <li ng:repeat="result in ctrl.myUploads.data track by result.data.id" class="upload-result">
+        <ui-image-editor image="result"></ui-image-editor>
+        <gr-delete-image class="flex-right"
+                         ng:if="ctrl.canBeDeleted(result)"
+                         images="[result]"></gr-delete-image>
+    </li>
+</ul>
